refactor: modify class structure
