+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,9 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
+            android:name=".view.KakaoMapActivity"
+            android:exported="false" />
+        <activity
             android:name=".view.SearchWindowActivity"
             android:exported="false" />
         <activity
