+++ app/src/main/res/layout/activity_main.xml
@@ -1,62 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.MainActivity">
-
-    <com.kakao.vectormap.MapView
-        android:id="@+id/map_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/goto_search_window"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="25dp"
-        android:background="@color/white"
-        android:elevation="10dp"
-        android:padding="15dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <TextView
-            android:id="@+id/goto_search_window_text"
-            android:layout_width="0dp"
-            android:layout_height="35dp"
-            android:gravity="center|left"
-            android:text="@string/searchWindowHint"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toLeftOf="@id/goto_search_window_img"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintWidth_default="spread" />
-
-        <ImageView
-            android:id="@+id/goto_search_window_img"
-            android:layout_width="25dp"
-            android:layout_height="25dp"
-            android:src="@drawable/ic_custom_magnifying_glass"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toRightOf="@+id/goto_search_window_text"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="parent"
-        >
-
-        <include
-            android:id="@+id/place_info_bottom_sheet"
-            layout="@layout/activity_place_info_bottom_sheet" />
-
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
+    tools:context=".view.KakaoMapActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
